refactor(gen:test): pull 4 helper functions out into a separate module
