+++ src/test/main.test.js
@@ -4,11 +4,15 @@ import fs from 'fs';
 import _ from 'lodash';
 import Promise from 'bluebird';
 Promise.promisifyAll(fs);
-import {exec} from 'child_process';
 import helpers from 'yeoman-test';
 import assert from 'yeoman-assert';
 import * as getExpectedFiles from './get-expected-files';
-import recursiveReadDir from 'recursive-readdir';
+import {
+  copyAsync,
+  runCmd,
+  assertOnlyFiles,
+  getConfig
+} from './test-helpers';
 
 const defaultOptions = {
   buildtool: 'grunt',
@@ -27,52 +31,6 @@ const defaultOptions = {
   socketio: true
 };
 const TEST_DIR = __dirname;
-
-function copyAsync(src, dest) {
-  return fs.readFileAsync(src)
-    .then(data => fs.writeFileAsync(dest, data));
-}
-
-/**
- * @callback doneCallback
- * @param {null|Error} err
- */
-
-/**
- * Run the given command in a child process
- * @param {string} cmd - command to run
- * @param {doneCallback} done
- */
-function runCmd(cmd, done) {
-  exec(cmd, {}, function(err, stdout, stderr) {
-    if(err) {
-      console.error(stdout);
-      throw new Error(`Error running command: ${cmd}`);
-      done(err);
-    } else {
-      if(DEBUG) console.log(stdout);
-      done();
-    }
-  });
-}
-
-function assertOnlyFiles(expectedFiles, topLevelPath='./', skip=['node_modules', 'bower_components']) {
-  return new Promise((resolve, reject) => {
-    recursiveReadDir(topLevelPath, skip, function(err, actualFiles) {
-      if(err) return reject(err);
-
-      actualFiles = _.map(actualFiles.concat(), file => path.normalize(file.replace(path.normalize(`${topLevelPath}/`), '')));
-      expectedFiles = _.map(expectedFiles, file => path.normalize(file));
-
-      let extras = _.pullAll(actualFiles, expectedFiles);
-
-      if(extras.length !== 0) {
-        return reject(extras);
-      }
-      resolve();
-    });
-  });
-}
 
 function runGen(prompts) {
   return new Promise((resolve, reject) => {
@@ -127,12 +85,6 @@ function runEndpointGen(name, opt={}) {
     gen
       .on('error', reject)
       .on('end', () => resolve())
-  });
-}
-
-function getConfig(dir) {
-  return fs.readFileAsync(path.join(dir, '.yo-rc.json'), 'utf8').then(data => {
-    return JSON.parse(data);
   });
 }
